backend/requirements.txt
# Web Framework
fastapi==0.110.1
uvicorn==0.25.0
python-multipart==0.0.20

# Database
motor==3.3.1
pymongo==4.5.0

# Authentication
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.3.0

# AI Service
google-genai==1.38.0

# File Processing
PyMuPDF==1.26.4
python-docx==1.2.0
Pillow==11.3.0

# PDF Generation
reportlab==4.4.3

# Environment and Configuration
python-dotenv==1.1.1

# Utilities
aiohttp==3.12.15
httpx==0.28.1
